// src/types.ts
export enum UserRole {
  MASTER = 'MASTER',
  OWNER = 'OWNER',
  ADMIN_OWNER = 'ADMIN_OWNER',
  ADMIN_DIVISI = 'ADMIN_DIVISI',
  MANAGER = 'MANAGER',
  KASIR = 'KASIR',
  WAREHOUSE = 'WAREHOUSE',
  STAFF = 'STAFF',
}
// ...
export interface OutletEntity {
  outletId: string;
  businessId: string;
  ownerId: string;
  name: string;
  code: string;
  location: string;
  phone?: string;
  assignedStaffId?: string;
  assignedStaffName?: string;
  status: 'ACTIVE' | 'INACTIVE';
  createdAt: number;
}